refactor: split common/tag.go — schema element creation into tag_schema.go
